+++ src/__tests__/assignment.test.ts
@@ -1,8 +1,8 @@
 import { describe, test } from '@jest/globals'
 
 import { testProgram } from '../interpreter/cInterpreter'
-import { INT_BASE_TYPE } from '../interpreter/typeUtils'
-import { intToBinary } from '../interpreter/utils'
+import { INT_BASE_TYPE } from '../interpreter/utils/typeUtils'
+import { intToBinary } from '../interpreter/utils/utils'
 import { expectLogOutputToBe, verifyProgramCompleted } from './utils'
 
 describe('assignments', () => {
